+++ packages/mobile/src/store/offline-downloads/sagas/downloadQueueSagas/watchAddOfflineItems.ts
@@ -1,12 +1,12 @@
 import { takeEvery, select, put } from 'typed-redux-saga'
 
-import { getDownloadQueue, getQueueStatus } from '../selectors'
+import { getDownloadQueue, getQueueStatus } from '../../selectors'
 import {
   addOfflineItems,
-  downloadQueuedItem,
+  requestDownloadQueuedItem,
   QueueStatus,
   updateQueueStatus
-} from '../slice'
+} from '../../slice'
 
 export function* watchAddOfflineItems() {
   yield* takeEvery(addOfflineItems.type, startQueueIfIdle)
@@ -18,6 +18,6 @@ function* startQueueIfIdle() {
 
   if (downloadQueue.length > 0 && queueStatus === QueueStatus.IDLE) {
     yield* put(updateQueueStatus({ queueStatus: QueueStatus.PROCESSING }))
-    yield* put(downloadQueuedItem())
+    yield* put(requestDownloadQueuedItem())
   }
 }
